==> 2. Scripts/1. Intro to Pandas.ipynb <==
# INTRO TO PANDAS

### This script contains the following:
#### 1. Importing libraries
#### 2. Turn folder path into a string
#### 3. Importing with os and vars_list
#### 4. Print first 5 rows of dataframe
#### 5. Importing with os
#### 6. head & tail functions
#### 7. Name of the columns
#### 8. Number of rows and columns
#### 9. Answers to key questions

# -------------------------------------------------------------------------------------------------------------

## 1. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Turn folder path into a string

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

## 3. Import orders.csv with os and vars_list

# To import a reduced version of the data, we create a list of the columns we need
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# We then give Python the instruction of reading the file located in that path
df = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'orders.csv'), usecols = vars_list, index_col = False)

## 4. Print first 5 rows of dataframe

# First five rows of dataframe
df.head()

## 5. Import products.csv with os

# To import the products.csv file, we follow the same path and give a name to the new dataframe
df_prods = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'products.csv'), index_col = False)

## 6. First 20 rows and last 30 rows from products data set

# To show the first 20 rows, we use the head function and establish the range
df_prods.head(20)

# To show the last 35 rows, we use the tail function and number the range.
df_prods.tail(35)

## 7. Name of the columns

df_prods.columns

## 8. Number of rows and columns

df_prods.shape

# We can see that our products dataframe has 49.693 rows and 5 columns

## 9. Answer the following questions

## 9.1. Max value of aisle_id column

# To determine the maximum value, we use the describe function which shows the descriptive statistics

df_prods.describe()

As we can see in the aisle_id column, max value is 134. 

## 9.2. Data type of department_id column

#For the data types we use the dtypes function

df_prods.dtypes

The department data type is int64.
