==> 2. Scripts/2. Importing data & Performing descriptive analysis.ipynb <==
# IMPORT DATA & DESCRIPTIVE ANALYSIS

### This script contains the following:
#### 1. Importing libraries
#### 2. Importing Data
#### 3. Exploring Data with pandas
#### 4. Python Data Structures & Lists


# ---------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Importing Data

df = pd.read_csv(r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis\2. Data\2.1. Original Data\orders.csv', index_col = False)

path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

path

df = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'orders.csv'), index_col = False)


## 3. Exploring Data with pandas

# First five rows of dataframe
df.head()

# First 30 rows of dataframe
df.head(30)

Last 5 rows of dataframe
df.tail()

# Dimensions of dataframe
df.shape

# Basic information about dataframe
df.info()

# Data types of the dataframe's columns
df.dtypes

# Names of the dataframe's columns
df.columns

# Descriptive statistics
df.describe()

## 4. Python Data Structures. Lists

# To import a reduced version of the data, firstly we create a list
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

# Then we import the data with the selected columns of the list
df = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'orders.csv'), usecols = vars_list)

df.head()
