==> 2. Scripts/3. Data wrangling and subsetting.ipynb <==
# DATA WRANGLING & SUBSETTING

### This script contains the following:
#### 1. Importing libraries
#### 2. Importing data
#### 3. Wrangling data
#### 4. Data dictionary
#### 5. Subsetting
#### 6. Exporting dataframes
#### *. TASK

# --------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Importing data

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

# We then give Python the instruction of reading the file located in that path
df_ords = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'orders.csv'), index_col = False)
df_prods = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'products.csv'), index_col = False)

## 3. Wrangling data

## 3.1. Dropping columns

# Firstly we will get rid of the columns that we don't need. In this case, we will drop "eval_set" from orders.csv
df_ords.drop(columns = ['eval_set'])

# Then we overwrite the previous dataframe with the new one, without the unwanted column.
df_ords = df_ords.drop(columns = ['eval_set'])

# We can look for missing values with the value_counts function
df_ords ['days_since_prior_order'].value_counts(dropna = False)

## 3.2. Renaming columns

# We can change column names to make them more representative.
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# We check if function worked as desired
df_ords.head()

## 3.3. Changing a variable's data type

# We can change the id variables to strings as we don't need statistical analysis on them.
df_ords['order_id'] = df_ords['order_id'].astype('str')

# We can double check that the change took place with dtype function. "0" means object/string 
df_ords ['order_id'].dtype

## 3.4. Transposing data

# We are going to start importing a new dataframe, departments
df_dep = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# To transpose it, the function is very simple.
df_dep.T

# Now we will create a new version of this vertical dataframe
df_dep_t = df_dep.T

# Let's see what it looks like
df_dep_t

## 3.4.1. Creating an index 

# We then create an index column with the rows' numbers.
df_dep_t.reset_index()

## 3.4.2. Creating a new header

# To create a new header, we need to follow 3 steps

# Step 1. Create the the header
# Firstly we take the first row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

new_header 

# We t Step 2. Get rid of old header
# Create a new dataframe with the info from row 1 onwards
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Step 3. We add the new header
# We use the list of our new header for the name of the columns
df_dep_t_new.columns = new_header

# This is how our new dataframe looks like now
df_dep_t_new

# 4. Creating a Data Dictionary

# We create the variable from our dataframe, after wrangling it to make it suitable to our needs.
data_dict = df_dep_t_new.to_dict('index')

data_dict

# To put it in use, we will bring the products file again into show
df_prods.head()

# To find out which department are these products, we would use the following command
print(data_dict.get('19'))

# For the rest of products in our head list
print(data_dict.get('13'))
print(data_dict.get('7'))
print(data_dict.get('1'))

# 5. Subsetting

# To create a subset of snacks
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# You could also use this command with the function "loc" 
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2.head()

# There is one more way, with the function "isin([])"
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3.head()

# 6. Exporting dataframes

# At the end of every notebook, we should be exporting our data to keep it safe and usable in other analysis.
# The command would take the following shape--> df_ords.to_csv(os.path.join(path, '2. Data','Prepared Data', 'orders_wrangled.csv'))

## --------------------------------------------------------------------------------------------------------------------------------

#                                                           TASK  

## 2. Find and make suitable a variable that doesn't need to be in numeric format 

# Firstly we represent the table - please note name change for next task point has already been applied in this representation
df_ords.head()

# User_id is a number to identify the customers. No statistical significance.
df_ords ['user_id'] = df_ords ['user_id'].astype('str')

# Order_number is also a way of identifying the different orders. No statistical significance.
df_ords ['order_number'] = df_ords ['order_number'].astype ('str')

## 3. Change the name of a variable with no intuitive name

# The last column seems to have a name a little bit confusing, I will change it to something more intuitive.
df_ords.rename(columns = {'days_since_prior_order' : 'days_from_last_order'}, inplace = True)

# Now we check the result applied...
df_ords.head()

## 4. Find the busiest time for placing orders

# To calculate the average we use the function mean
df_ords['order_hour_of_day'].mean()

# The average of 13.45 tell us that the busiest hour for orders is at some point between 1.25pm and 1.30pm

## 5. Determine the value of department_id number 4 with a Data Dictionary

# Firstly we represent the dictionary we created before, it will come handy the next exercises
data_dict

# To find out the name of each department, in this case number 4, we would use the following command
print(data_dict.get('4'))

## 6. Create a subset of breakfast items for the sales department

# To create a subset of breakfast items, I checked in the dictionary which number of department id refers to breakfast (14)
df_breakfast = df_prods.loc [df_prods['department_id']==14]

# We show a representation...
df_breakfast.head()

## 7. Create a subset with items suitable for dinner parties

# We want to create a dataframe representing alcohol, beverages, deli and meat&seafood
df_dinner_parties = df_prods.loc [df_prods['department_id'].isin([5,7,12,20])]

# Let's have a look at our new dataframe
df_dinner_parties

## 8. Track the count of rows of the last dataframe

# As we can see two lines above, the dinner_parties df has 7650 rows

## 9. Extract all the information you can from the user 1

# In other words, we have been asked to create a subset for user_id 1. We must put '1' because we changed the data type to string.
df_user_1 = df_ords.loc[df_ords['user_id'].isin(['1'])]

df_user_1

# I would also show the statistical analysis, as well as some basic info of the dataset
df_user_1.describe()

# The info doesn't tell us much
df_user_1.info()

## 10. Provide some details about this customer's behaviour

# From the statistical analysis we can throw some findings.
  # Customer has order 11 times, always between Monday and Thursday (min=1, max=4)
  # He always orders between 7am and 4pm, being around 10am o'clock the average time
  # He orders every 19 days as an average. However, in one ocassion he ordered the same day twice, with two hours
  # of difference, and after that he was 30 days without ordering at all, his longest gap.

## 12. Export your data frames

# Firstly we will export df_ords
df_ords.to_csv(os.path.join(path, '2. Data','2.2. Prepared Data', 'orders_wrangled.csv'))

# Then we export the one we wrangled for the departments
df_dep_t_new.to_csv(os.path.join(path, '2. Data', '2.2. Prepared Data', 'departments_wrangled.csv'))
