==> 2. Scripts/4. Data Consistency Checks.ipynb <==
# DATA CONSISTENCY CHECKS

### This script contains the following:
#### 1. Importing libraries
#### 2. Importing data
#### 3. Data consistency checks
#### 4. Exporting dataframes
#### *. TASK

# ---------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Importing data

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

# We then give Python the instruction of reading the file located in that path
df_ords = pd.read_csv(os.path.join(path, '2. Data', '2.2. Prepared Data', 'orders_wrangled.csv'), index_col = False)
df_prods = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'products.csv'), index_col = False)

## 3. Data consistency checks

# We can start with a statistical analysis
df_ords.describe()

# In it, we check if the maximums and minimums make sense, if percentil 50 in dow is 3, and so.

## 3.1. Mixed-type data

# For the sake of learning, we are going to create a dataframe
df_test = pd.DataFrame()

# Now we create a mixed-type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# The formula to check for myxed_type columns

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# After checking, you have to fix the issue by converting the column in the most common type
df_test['mix'] = df_test['mix'].astype('str')

## 3.2. Missing values

# To find missing values, we use the following command
df_prods.isnull().sum()

# To view a dataframe showing the missing values only
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Because the missing values are names, we can't impute anything. We will create a new df with the non-missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Let's compare the size of both dataframes to see if our changed applied correctly 
df_prods.shape

df_prods_clean.shape

## 3.3. Duplicates

# To find duplicates
df_dups = df_prods_clean [df_prods_clean.duplicated()]

df_dups

# To create a new df, getting rid of the duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# To check if the changes applied properly
df_prods_clean_no_dups.shape

## 3.4. Incoherent data

# For prices over $100, we will put a mark as a missing value as they can mess up our analysis. 
df_prods_clean_no_dups.loc[df_prods_clean_no_dups['prices'] >100, 'prices'] = np.nan

# To check our result was applied
ords_prods_merge['prices'].max()

## 4. Exporting our dataframes

# To finalise, we export our dataframe clean of missing values and duplicates
df_prods_clean_no_dups.to_csv(os.path.join(path, '2. Data', '2.2. Prepared Data', 'prods clean no dups.csv'))

# ------------------------------------------------------------------------------------------------------

# TASK 4.5

## 2. Interpret the statistical analysis of prods

df_prods_clean_no_dups.describe()

From these results we can deduct that there is a mistake in the pricing of at least one of the products, since the Max price is $99999. I will conduct my own investigation to find out more about it

# To identify rows with a wrong price, I chose a very high number 
df_wrong_price = df_prods_clean_no_dups[df_prods_clean_no_dups['prices']>50000]

df_wrong_price

 We can see that there is only one row with this issue. Potentially we would be thinking of getting rid of this row

## 3. Check for myxed-type data in the orders dataframe

# To check for myxed-type data

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

## 4. If you find myxed-type data, fix it

There is no myxed-type data, every column is consistent in terms of data types

## 5. Check for missing values in the orders dataframe

# To find out missing values
df_ords.isnull().sum()

# To view a dataframe showing the missing values only
df_ords_nan = df_ords[df_ords['days_from_last_order'].isnull() == True]

df_ords_nan.head()

df_ords_nan.tail()

At a quick glance we can see that every first order for every customer is a missing value in the days_from_last_order, something obvious since they have never ordered before.

## 6. Address the missing values

In this case we cannot impute mean or median because we would be introducing wrong information. We can't eliminate the missing data rows either as they are meaningful, they mean the first orders for each customer. In this case it would be more suitable to create a new variable that act like a flag based on the missing values. We are going to call this variable "first_order" and it will show whether an order is the first that the associated specific customer has made. 

# To add the first_order column to our dataframe, we use the following command
df_ords ['first_order'] = df_ords['days_from_last_order'].isnull() == True

# To check the result
df_ords.head()

Our missing values take now much more sense

## 7. Check for duplicates values in the orders dataframe

# To find the duplicates we use the following command
df_ords_dups = df_ords [df_ords.duplicated()]

# Let's see how many we have
df_ords_dups

As we can see, we don't have any duplicates in the orders dataframe.

## 8. Address the duplicates

There is no need as there is no duplicates

## 9. Export the cleaned data

# First of all,I am going to rename my orders dataframe 
df_ords_checked = df_ords

# Now we are going to export them to our 'Prepared Data' folder
df_ords_checked.to_csv(os.path.join(path, '2. Data','2.2. Prepared Data', 'orders_checked.csv'))

# We export now the products dataframe
df_prods_clean_no_dups.to_csv(os.path.join(path, '2. Data', '2.2. Prepared Data', 'products_checked.csv'))
