==> 2. Scripts/5. Combining & Exporting Data (Part 1).ipynb <==
# COMBINING & EXPORTING DATA

### This script contains the following:
#### 1. Importing  libraries
#### 2. Creating a path
#### 3. Import the orders product combined dataframe
#### 4. Shape
#### 5. Combining dataframes
#### 6. Merge flag
#### 7. Exporting dataframes

# --------------------------------------------------------------------------------------------------------------- 

### 1. Importing the libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 2. Creating a path

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

### 3. Import the orders product combined dataframe

# We then give Python the instruction of reading the file located in that path
df_ords_prods_comb = pd.read_pickle(os.path.join(path, '2. Data', '2.2. Prepared Data', 'orders_products_combined.pkl'))

# We will also import the products file
df_prods_check = pd.read_csv(os.path.join(path, '2. Data', '2.2. Prepared Data', 'products_checked.csv'), index_col = False)

### 4. Check the shape 

# We check the dimensions of our newly imported dataframe
df_ords_prods_comb.shape

# Check the appearance
df_ords_prods_comb.head()

# We also check the products file
df_prods_check.shape

df_prods_check.head()

# Get rid of the first column, not representing anything
df_prods_check = df_prods_check.drop (columns = ['Unnamed: 0'])

df_prods_check.head()

### 5. Determine a suitable way to combine them

# In this scenario, we would want to keep all the information in both dataframes
df_ords_prods_merged = df_ords_prods_comb.merge(df_prods_check, on = 'product_id')

# We check the output
df_ords_prods_merged.head()

### 6. Confirm the results with the merge flag

# We use the value counts to see the merge results
df_ords_prods_merged['_merge'].value_counts()

### 7. Export the newly created dataframe

# Export data to pkl

df_ords_prods_merged.to_pickle(os.path.join(path, '2. Data','2.2. Prepared Data', 'orders_products_merged.pkl'))
