==> 2. Scripts/6. Merging Instacart data (Part 2).ipynb <==
# MERGING INSTACART DATA

### This script contains the following:
#### 1. Importing libraries
#### 2. Importing data
#### 3. Merging data
#### 4. Exporting data

# --------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Importing data

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

# We then give Python the instruction of reading the file located in that path
df_ords = pd.read_csv(os.path.join(path, '2. Data', '2.2. Prepared Data', 'orders_checked.csv'), index_col = False)
df_ords_prior = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'order_products__prior.csv'), index_col = False)

## 3. Merging data

# We check the dimensions of our newly imported dataframe
df_ords_prior.shape

# Check the dataframe has no apparent issues
df_ords_prior.head()

# Check the other df
df_ords.shape

# Check its appearance
df_ords.head()

# Get rid of the first two columns, not representing anything
df_ords = df_ords.drop (columns = ['Unnamed: 0', 'Unnamed: 0.1'])

# Check dimensions again
df_ords.shape

# We then combine both data sets
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check the output
df_merged_large.head()

# We sum up the values to see if they are present in both dataframes or just one
df_merged_large['_merge'].value_counts()

## 4. Exporting data

# Export to pickle
df_merged_large.to_pickle(os.path.join(path, '2. Data','2.2. Prepared Data', 'orders_products_combined.pkl'))

# Check the appearance
df_merged_large.head()

df_merged_large.shape
