==> 2. Scripts/7. Deriving new variables .ipynb <==
# DERIVING NEW VARIABLES

### This script contains the following:
#### 1. Importing libraries
#### 2. Importing data
#### 3. If-statements
#### *. TASK

# --------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Importing data

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

# We then give Python the instruction of reading the file located in that path
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '2. Data', '2.2. Prepared Data', 'orders_products_merged.pkl'))

## 3. If-statements

### 3.1. If-statements with User-defined functions

# Firstly, we will choose a subset of 1M rows, since the data is very large
df_subset = df_ords_prods_merged [:1000000]

# Check dimensions
df_subset.shape

# We define our new variable and give certain conditions 
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Now, to see the result of the function, we add a column to our df
df_subset['price_range'] = df_subset.apply(price_label, axis=1)

# To check the values in our new column
df_subset['price_range'].value_counts(dropna = False)

# There is no high range products so let's find out which product in the subset has the highest price
df_subset['prices'].max()

### 3.2. If-statements with the loc() function

# Creating the conditions. For loc() function, we will be using one row per each function
df_subset.loc[(df_subset['prices'] <= 15) & (df_subset['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_subset.loc[df_subset ['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_subset.loc[df_subset['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_subset['price_range_loc'].value_counts(dropna=False)

# Let's try to do it with our merged data
df_ords_prods_merged.loc[df_ords_prods_merged ['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods_merged.loc[df_ords_prods_merged['prices']> 15, 'price_range_loc'] = 'High range product'

df_ords_prods_merged['price_range_loc'].value_counts(dropna = False)

### 3.3. If-statements with For-loops

df_ords_prods_merged.head()

# Creating a column to summarise how busy each day of the week is
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna=False)

#### Each number represent a day. From the project we know that Saturday = 0, meaning this is the busiest day of the week

# Creating a column that displays the busiest, the least busy and regularly busy days of the weak
result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Let's see the output
result

# To add the column to the dataframe
df_ords_prods_merged ['busiest_day'] = result

df_ords_prods_merged ['busiest_day'].value_counts(dropna = False)

#   

# --------------------------------------------------------------------------------------------------------

# 

# TASK 7

### 2. Create a "Busiest days" and "Less busy days" including two days in each list

# Creating a column that displays the 2 busiest, the 2 less busy and regularly busy days of the week
result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0 or value == 1 :
    result.append("Two_busiest_days")
  elif value == 3 or value == 4 :
    result.append("Two_slowest_days")
  else:
    result.append("Normal_days")

result

# To add the column to the dataframe
df_ords_prods_merged ['Busiest_days'] = result

### 3. Check the values in this new column

df_ords_prods_merged ['Busiest_days'].value_counts(dropna = False)

#### We can see that Sat & Sun have almost as many orders as Mon, Thu & Fri together. Tue & Wed are the quietest days.

### 4. Create a new column for the busiest period of day

# We create a column to summarise how busy each hour is
df_ords_prods_merged['order_hour_of_day'].value_counts(dropna=False)

# Creating a column that displays the hours with the most orders, the least orders and the hours with average orders.
result = []

for value in df_ords_prods_merged["order_hour_of_day"]:
  if value == 3 or value == 4 or value == 2 or value == 5:
    result.append("Less busy hours")
  elif value == 1 or value ==0 or value ==6 or value ==23 or value ==22 or value ==21 or value ==7 or value ==20:
    result.append("Average busy hours")
  else:
    result.append("Most busy hours")

# To add the column to the dataframe
df_ords_prods_merged ['busiest_period_of_day'] = result

# To check the output
result

### 5. Check the frequency for this new column

df_ords_prods_merged ['busiest_period_of_day'].value_counts (dropna = False)

#### Most orders occur between 10am-3pm. 
#### From 12 am-5am the amount of orders is quite small, thy will need less staff.
#### 5am-10am & 3pm-12am are considered periods with an average number of orders

### 

### 7. Export your dataframe as a pickle file

# To export the df as a pkl
df_ords_prods_merged.to_pickle(os.path.join(path, '2. Data','2.2. Prepared Data', 'orders_products_merged_new_variables.pkl'))
