==> 2. Scripts/8.2. Grouping Data & Aggregating variables (Part 2).ipynb <==
# GROUPING DATA & AGGREGATING VARIABLES (part 2)


### This script contains the following:
#### 1. Importing data and libraries
#### 2. Aggregate mean of order_number by department_id
#### 3. Analyze the results
#### 4. Loyalty flag
#### 5. Basic statistics of product prices
#### 6. Spending flag
#### 7. Order frequency flag

# ---------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

### 1.1. Importing data

# We create a path to add a shortcut to our files location
path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\Data Immersion\Python\Instacart Basket Analysis'

# We then give Python the instruction of reading the file located in that path
ords_prods_merge = pd.read_pickle(os.path.join(path, '2. Data', '2.2. Prepared Data', 'orders_products_merged_new_variables.pkl'))

## 2. Find aggregate mean of order_number by department_id

# To calculate the average number of orders per department ID
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

## 3. Analyze the results

### When comparing the results between a subset and the entire dataframe, we notice that the averages of orders for all of the departments change. For example, the department 17 had an average of 11.29 in the subset but increases to 15.69 when using the whole dataframe as a reference.

## 4. Create a loyalty flag

# Firstly we create a new column, group by criteria and apply the function
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# We check the output 
ords_prods_merge.head(15)

# With the new column ready, we have to create a flag to denominate the loyalty of the customers
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# We do the same for regular customers
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

# And for new customers
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# As a good practice, we check the frequency 
ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

 ## 5. Check the basic statistics of the product prices for each loyalty category

# We are going to calculate average, minimum and maximum of the prices for each category of customer
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

### As we can see, and against my expectations, it is the new customers who spent more in each product as an average, followed by the regular customers ,being the loyal customers the ones that spends the less in average. That might be due to the fact that they know about offers or they get discount, e.g. 

## 6. Create a spending flag for each user based on the average price across all their orders

# We create a new column, group by criteria and apply the function
ords_prods_merge['average_spending'] = ords_prods_merge.groupby('user_id').agg({'prices':['mean']})

# CHeck the ouput 
ords_prods_merge.head()

# With the new column ready, we have to create a flag to denominate their spending category
ords_prods_merge.loc[ords_prods_merge['average_spending'] < 10 , 'spending_flag'] = 'Low spender'

# We do the same for high spender
ords_prods_merge.loc[ords_prods_merge['average_spending'] >= 10, 'spending_flag'] = 'High spender'

# We check the frequency
ords_prods_merge ['spending_flag'].value_counts(dropna = False)

## 7. Create an order frequency flag based on the regularity of an user's ordering behaviour

# We create a new column, group by criteria and apply the function
ords_prods_merge['frequency_orders'] = ords_prods_merge.groupby('user_id').agg({'days_from_last_order':['median']})

# Check the output
ords_prods_merge.head()

# With the new column ready, we have to create a flag to denominate the customers based on their spending behaviour
ords_prods_merge.loc[ords_prods_merge['frequency_orders'] > 20, 'frequency_flag'] = 'Non-frequent customer'

# We do the same for regular customers
ords_prods_merge.loc[(ords_prods_merge['frequency_orders'] > 10) & (ords_prods_merge ['frequency_orders'] <=20), 'frequency_flag'] = 'Regular customer'

# And for the frequent customers
ords_prods_merge.loc[ords_prods_merge['frequency_orders'] <= 10, 'frequency_flag'] = 'Frequent customer'

# We check the results
ords_prods_merge ['frequency_flag'].value_counts(dropna = False)

# To export the df as pickle
ords_prods_merge.to_pickle(os.path.join(path, '2. Data','2.2. Prepared Data', 'orders_products_merged_with_aggregations.pkl'))
