==> 2. Scripts/9.1. Intro to Data Visualizations with Python (Part 1).ipynb <==
# DATA VISUALIZATIONS WITH PYTHON (part 1)

### This script contains the following:
#### 1. Importing data and libraries
#### 2. Checking and wrangling data
#### 3. Data quality and consistency checks
#### 4. Combining data sets
#### 5. Exporting data sets

# --------------------------------------------------------------------------------------------------------------

## 1. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

### 1.1. Importing data

# We create a path to add a shortcut to our files location

path = r'C:\Users\javis\OneDrive\Documentos\Career Foundry\2. Data Immersion\4. Python\Instacart Basket Analysis'

# We then give Python the instruction of reading the file located in that path

df_custom = pd.read_csv(os.path.join(path, '2. Data', '2.1. Original Data', 'customers.csv'),index_col = False)

## 2. Checking and wrangling data

# We check the dimensions

df_custom.shape

# Check the types of the columns

df_custom.dtypes

# Check the look

df_custom.head()

# Change data type of user_id as we don't need analysis on them

df_custom['user_id'] = df_custom['user_id'].astype('str')

# We rename the columns to have everything lower case

df_custom= df_custom.rename(columns = {'First Name':'first_name', 'Surnam':'surname', 'Gender':'gender', 'STATE':'state', 'Age':'age','n_dependants':'dependants'})

# We check for missing values

df_custom.isnull().sum()

# To avoid errors later, we are going to drop the name column

df_custom.drop(columns = ['first_name'])

# We overwrite the previous dataframe

df_custom = df_custom.drop(columns = ['first_name'])

## 3. Data quality and consistency checks

# We look for duplicates

df_custom_dups = df_custom [df_custom.duplicated()]

# We drop the duplicates 

df_custom = df_custom.drop_duplicates()

# We check for mixed data types
for col in df_custom.columns.tolist():
    weird = (df_custom[[col]].applymap(type) != df_custom[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_custom[weird]) > 0:
        print (col)

# Lets check dimensions now to see if anything changed

df_custom.shape

# Check the appearance

df_custom.head()

### 3.1. Exporting file 

# Exporting file as pickle

df_custom.to_pickle(os.path.join(path,'2. Data', '2.2. Prepared Data', 'customers_checked.pkl'))

## 4. Combining data sets¶

# To combine the datasets, firstly we must import the another one

df_ords_prods = pd.read_pickle(os.path.join(path, '2. Data', '2.2. Prepared Data', 'orders_products_merged.pkl'))

# To check 'user_id' data type in customers dataframe

df_custom ['user_id'].dtype

# To check 'user_id' data type in our previous dataframe

df_ords_prods ['user_id'].dtype

# We need to change the data type in the orders set so they have same type and we can combine them.
# We will change it to string

df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('str')

# To merge the data sets

df_cust_ord_prods = df_ords_prods.merge(df_custom, on = 'user_id')

## 5. Exporting data sets

# To export the combined data sets as a pickle file

df_cust_ord_prods.to_pickle(os.path.join(path,'2. Data', '2.2. Prepared Data', 'cust_orders_products_combined.pkl'))
